clean up http routing in main.cpp
